SF-ID: 2949655: fixed casting exception that was breaking bomSimulations
git-svn-id: svn://svn.opentaps.org/opentaps_all/versions/1.4/trunk@14957 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/opentaps-common/src/common/org/opentaps/domain/manufacturing/bom/BomNode.java
@@ -564,9 +564,9 @@ public class BomNode implements BomNodeInterface {
             Map inputContext = UtilMisc.toMap("arguments", arguments, "userLogin", userLogin);
             try {
                 Map result = dispatcher.runSync(serviceName, inputContext);
-                Double calcQuantity = (Double) result.get("quantity");
+                BigDecimal calcQuantity = (BigDecimal) result.get("quantity");
                 if (calcQuantity != null) {
-                    this.quantity = BigDecimal.valueOf(calcQuantity);
+                    this.quantity = calcQuantity;
                 }
             } catch (GenericServiceException e) {
                 Debug.logError(e, "Problem calling the getManufacturingComponents service", MODULE);
